+++ client/src/components/UserSearchResult/index.tsx
@@ -5,8 +5,8 @@ import FollowerUserContextMenu from '@components/FollowerUserContextMenu';
 import UserItem from '@components/UserItem';
 import UserList from '@components/UserList';
 import { EllipsisHorizontalIcon } from '@heroicons/react/20/solid';
-import { useFollowersMapQuery } from '@hooks/useFollowersQuery';
 import { useFollowingsMapQuery } from '@hooks/useFollowingsQuery';
+import { useFollowersMapQuery } from '@hooks/user';
 import { useRootStore } from '@stores/rootStore';
 import React from 'react';
 import Scrollbars from 'react-custom-scrollbars-2';
